backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/LLM
litellm==1.10.0
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0

# Async HTTP
aiohttp==3.9.1
aiosignal==1.3.1

# Web Scraping
newspaper4k==0.9.0
lxml[html_clean]==4.9.3
beautifulsoup4==4.12.2
feedparser==6.0.10

# Task Queue
celery==5.3.4
redis==5.0.1

# Feishu/Lark
lark-oapi==1.0.0

# Email
aiosmtplib==3.0.1

# Utils
python-dateutil==2.8.2
python-multipart==0.0.6
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
